Fix compiler warnings that are coming Aria production code:
Warning: Aria.SDK.CPP\Solutions\Clienttelemetry\..\..\lib\pal\PAL_Win32.cpp(280): warning C4459: declaration of 'ticksPerSecond' hides global declaration
Change: zlib has a global ticksPerSecond. Renamed this one to ticksInOneSecond

Warning: Aria.SDK.CPP\lib\pal\win32\WindowsDesktopDeviceInformationImpl.cpp(62): warning C4996: '_strlwr': This function or variable may be unsafe. Consider using  _strlwr_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS.
See online help for details.
Change: Added lstrlen to get length. In this case, it's probably not needed, but this code is called infrequently, so the perf hit is negligible

Warning: Aria.SDK.CPP\lib\pal\win32\NetworkDetector.cpp(489): warning C4245: 'argument': conversion from 'int' to 'UINT', signed/unsigned mismatch
Change: Explicitly cast the -1 to a UINT

+++ lib/pal/win32/NetworkDetector.cpp
@@ -486,7 +486,7 @@ namespace Microsoft {
                 /// </summary>
                 void NetworkDetector::Stop()
                 {
-                    PostThreadMessage(m_listener_tid, -1, 0, NULL);
+                    PostThreadMessage(m_listener_tid, (UINT)-1, 0, NULL);
                     if ( isRunning && netDetectThread.joinable() )
                     {
                         std::unique_lock<std::mutex> lock(m_lock);
